requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
opencv-python==4.8.0.76
Pillow==9.5.0
dlib==19.24.2
scipy==1.10.1